// desktop/experimental/ui/src/types/rest.ts
export type ServerLegal = {
  legal_version: string
  media_retention_days: number
  operator_country: string
  operator_email: string
  operator_name: string
  project_notice: string
  session_expiry_days: number
  stun_servers: Array<string>
  turn_servers: Array<string>
  update_check: boolean
}